Restructure to work with esModuleInterop: true
