+++ test/types/pino-type-only.test-d.ts
@@ -3,6 +3,8 @@ import { expectType } from "tsd";
 import pino from "../../";
 import type {Logger, LogFn, P} from "../../pino";
 
+// NB: can also use `import * as pino`, but that form is callable as `pino()`
+// under `esModuleInterop: false` or `pino.default()` under `esModuleInterop: true`.
 const log = pino();
 expectType<Logger>(log);
 expectType<LogFn>(log.info);
